move universe selection to a separate class
