+++ src/machine_learning.py
@@ -11,97 +11,11 @@
 # %reload_ext autoreload
 # %autoreload 2
 
-# Load base and 3rd party packages
 import pandas as pd
 import numpy as np
 import matplotlib.pyplot as plt
 import statsmodels.api as sm
 from optimization_data import OptimizationData
-import onnx
-from abc import ABC, abstractmethod
-import tensorflow as tf
-
-#### Calculate the metrics RMSE and MAPE ####
-def calculate_rmse(y_true, y_pred):
-    """
-    Calculate the Root Mean Squared Error (RMSE)
-    """
-    rmse = np.sqrt(np.mean((np.array(y_true) - np.array(y_pred.values)) ** 2))
-    return rmse
-
-
-def calculate_mape(y_true, y_pred):
-    """
-    Calculate the Mean Absolute Percentage Error (MAPE) %
-    """
-    y_pred, y_true = np.array(y_pred), np.array(y_true)
-    mape = np.mean(np.abs((y_true - y_pred) / y_true)) * 100
-    return mape
-
-def show_result(predictions, y_test, y_actual, method = None):
-    print(f'RMSE of linear regression: {calculate_rmse(y_test, predictions)}')
-    print(f'MAPE of linear regression: {calculate_mape(y_test, predictions)}')
-
-    plt.plot(y_actual, color = 'cyan')
-    plt.plot(predictions, color = 'green')
-    plt.legend(["True values", "Prediction"])
-    plt.title(method)
-    plt.show()
-class Universe_selection(ABC):
-    def __init__(self,
-                 model: str,
-                 data: pd.DataFrame):
-        self.model = model
-        self.data = data
-
-    @abstractmethod
-    # def build_model(self):
-    #     self.trained_model = None
-    #     return None
-
-    def load_model(self, model_path: str = None):
-        self.trained_model = onnx.load(model_path)
-        print(onnx.checker.check_model(self.trained_model))
-
-    def select_universe(self, X, nb_stocks = 20):
-        self.trained_model.predict(X)
-        return None
-
-class LSTM_revelance(Universe_selection):
-
-    def load_model(self, model_path: str = "../model/lstm_msci_00.keras"):
-        self.trained_model = tf.keras.models.load_model(model_path)
-        # Show the model architecture
-        self.trained_model .summary()
-
-    def select_universe(self, X, nb_stocks = 20):
-        result_loaded_model = self.trained_model(X)
-        indx_top = result_loaded_model[-1].numpy().argsort()[:nb_stocks]
-        return self.data.columns[indx_top] # Name/code of stocks
-
-def train_test_split(X, y = None, queryonnx_model_path = None, test_size = 0.2) :
-    nb_test = int(test_size * len(X))
-    test_index = X.index[-nb_test:]
-    train_index = X.index[:X.shape[0] - nb_test]
-
-    X_test  = X.loc[test_index]
-    X_train = X.loc[train_index]
-
-    if not(y is None):
-        y_test  = y.loc[test_index]
-        y_train = y.loc[train_index]
-    else:
-        y_test, y_train = None, None
-
-    if not(query is None):
-        query_test = query.loc[test_index]
-        query_train = query.loc[train_index]
-    else:
-        query_test = None
-        query_train = None
-    return X_train, X_test, y_train, y_test, query_train, query_test
-
-
 
 
 if __name__ == '__main__':
@@ -124,7 +38,6 @@ if __name__ == '__main__':
     X.fillna(0, inplace=True)
 
     training_data = OptimizationData(align = True, lags = {'y': -1}, X = X, y = y_monthly)
-
 
     # Using statsmodels
